Moved calculation of iface_settings from CIDR to helper function
Also added a bit to use the first IP of the subnet as GW if no
other value is supplied. If default_gw is set True on one iface
(or set False on all other ifaces) this one will be used for
the default route.

There's still a problem with OVS-Bridges though...